Edit responsiveness and video in create post
changed the component of search in navBar to add search recommendations.
improved responsiveness of the navBar as the icons will be changed in the mobile mode with dotes to be able to use the search box.
adding a video in create post is fixed in addition to drag and drop.